requirements.txt
pdf2image==1.17.0
pillow==10.4.0
PyPDF2==3.0.1
tk==0.1.0
typing_extensions==4.12.2
